fixed:stimulusのタグ表示処理などのリファクタリング、まちレポ作成画面の修正
